All tests pass

+++ MaxRunSoftware.Utilities.Ftp.Tests/FtpClientTests.cs
@@ -28,33 +28,44 @@ public abstract class FtpClientTests<T> : TestBase, IDisposable where T : FtpCli
         Client = CreateClient();
     }
 
+    public void Dispose()
+    {
+        var c = Client;
+        Client = null!;
+        // ReSharper disable once ConditionalAccessQualifierIsNonNullableAccordingToAPIContract
+        c?.Dispose();
+    }
+
     protected abstract T CreateClient();
 
 
 
     [SkippableFact]
-    public void WorkingDirectory()
-    {
-        var o = Client.WorkingDirectory;
-        Assert.NotNull(o);
-        output.WriteLine(nameof(Client.WorkingDirectory) + ": " + o);
-    }
+    public void WorkingDirectory() => log.LogInformationMethod(new(), "{WorkingDirectory}", Client.WorkingDirectory);
 
     [SkippableFact]
-    public void ServerInfo()
-    {
-        var o = Client.ServerInfo;
-        output.WriteLine(nameof(Client.ServerInfo) + ": " + o);
-    }
+    public void WorkingDirectory_Not_Null() => Assert.NotNull(Client.WorkingDirectory);
+
+    [SkippableFact]
+    public void WorkingDirectory_Starts_With_DirectorySeparator() => Assert.StartsWith(Client.DirectorySeparator, Client.WorkingDirectory);
+
+    [SkippableFact]
+    public void WorkingDirectory_Does_Not_End_With_DirectorySeparator() => Assert.False(Client.WorkingDirectory.EndsWith(Client.DirectorySeparator));
+
+    [SkippableFact]
+    public void ServerInfo() => log.LogInformationMethod(new(), "{ServerInfo}", Client.ServerInfo);
 
     [SkippableFact]
     public void PutFile()
     {
+        var _ = CreateTestWorkingDirectory();
+        Assert.Empty(GetFiles());
+
         HelperDeleteFile("myfile.txt");
-        var file = HelperGetFile("myfile.txt");
+        var file = GetFile("myfile.txt");
         Assert.Null(file);
-        Client.PutFile("myfile.txt", Constant.Encoding_UTF8.GetBytes("Hello World"));
-        file = HelperGetFile("myfile.txt");
+        Client.PutFile("myfile.txt", ToBytes("Hello World"));
+        file = GetFile("myfile.txt");
         output.WriteLine(nameof(Client.PutFile) + ": " + file);
         Assert.NotNull(file);
     }
@@ -62,13 +73,14 @@ public abstract class FtpClientTests<T> : TestBase, IDisposable where T : FtpCli
     [SkippableFact]
     public void PutFile_GetFile()
     {
-        HelperDeleteFile("myfile.txt");
-        var file = HelperGetFile("myfile.txt");
+        CreateTestWorkingDirectory();
+
+        var file = GetFile("myfile.txt");
         Assert.Null(file);
         var fileOldString = "Hello World";
-        var fileOldBytes = Constant.Encoding_UTF8.GetBytes(fileOldString);
+        var fileOldBytes = ToBytes(fileOldString);
         Client.PutFile("myfile.txt", fileOldBytes);
-        file = HelperGetFile("myfile.txt");
+        file = GetFile("myfile.txt");
         Assert.NotNull(file);
         output.WriteLine(nameof(Client.PutFile) + ": " + file);
 
@@ -77,66 +89,93 @@ public abstract class FtpClientTests<T> : TestBase, IDisposable where T : FtpCli
         Assert.Equal(fileOldBytes.Length, fileNewBytes.Length);
         Assert.True(fileOldBytes.EqualsBytes(fileNewBytes));
 
-        var fileNewString = Constant.Encoding_UTF8.GetString(fileNewBytes);
+        var fileNewString = ToString(fileNewBytes);
         Assert.Equal(fileOldString, fileNewString);
     }
-
-    protected static byte[] ToBytes(string str) => Constant.Encoding_UTF8.GetBytes(str);
-    protected static string ToString(byte[] bytes) => Constant.Encoding_UTF8.GetString(bytes);
 
 
     [SkippableFact]
     public void PutFile_Auto_Create_SubDir_Multiple()
     {
         var dirName = CallerInfo.GetName()!;
-        var dir = HelperGetDirectory(dirName);
-        if (dir != null) Client.DeleteDirectory(dirName);
-        dir = HelperGetDirectory(dirName);
+        var dir = GetDirectory(dirName);
+        if (dir != null)
+        {
+            Client.DeleteDirectory(dirName);
+            dir = GetDirectory(dirName);
+        }
         Assert.Null(dir);
 
-        Client.PutFile($"{dirName}/1/myfile1.txt", ToBytes("My File 1"));
-        Client.PutFile($"{dirName}/1/2/myfile2.txt", ToBytes("My File 2"));
-        Client.PutFile($"{dirName}/1/2/3/myfile3.txt", ToBytes("My File 3"));
+        Client.PutFile(ConvertPath($"{dirName}/1/myfile1.txt"), ToBytes("My File 1"));
+        Client.PutFile(ConvertPath($"{dirName}/1/2/myfile2.txt"), ToBytes("My File 2"));
+        Client.PutFile(ConvertPath($"{dirName}/1/2/3/myfile3.txt"), ToBytes("My File 3"));
 
-        Client.PutFile($"{dirName}/4/5/6/myfile6.txt", ToBytes("My File 6"));
-        var files = HelperGetFilesRecursive(dirName);
+        Client.PutFile(ConvertPath($"{dirName}/4/5/6/myfile6.txt"), ToBytes("My File 6"));
+        var files = GetFilesRecursive(dirName);
 
         var wd = Client.WorkingDirectory;
         Assert.Equal(4, files.Length);
-        Assert.Contains(files, o => o.FullName == $"{wd}/{dirName}/1/myfile1.txt");
-        Assert.Contains(files, o => o.FullName == $"{wd}/{dirName}/1/2/myfile2.txt");
-        Assert.Contains(files, o => o.FullName == $"{wd}/{dirName}/1/2/3/myfile3.txt");
-        Assert.Contains(files, o => o.FullName == $"{wd}/{dirName}/4/5/6/myfile6.txt");
+        Assert.Contains(files, o => o.NameFull == ConvertPath($"{wd}/{dirName}/1/myfile1.txt"));
+        Assert.Contains(files, o => o.NameFull == ConvertPath($"{wd}/{dirName}/1/2/myfile2.txt"));
+        Assert.Contains(files, o => o.NameFull == ConvertPath($"{wd}/{dirName}/1/2/3/myfile3.txt"));
+        Assert.Contains(files, o => o.NameFull == ConvertPath($"{wd}/{dirName}/4/5/6/myfile6.txt"));
+    }
+
+    [SkippableFact]
+    public void WorkingDirectory_Change_To_Dot()
+    {
+        var wd1 = Client.WorkingDirectory;
+        log.LogInformationMethod(new(), "wd1: {WorkingDirectory}", wd1);
+
+        Client.WorkingDirectory = ".";
+        var wd2 = Client.WorkingDirectory;
+        log.LogInformationMethod(new(), "wd2: {WorkingDirectory}", wd2);
+        Assert.Equal(wd1, wd2);
+
+        Client.WorkingDirectory = ConvertPath("././././.");
+        var wd3 = Client.WorkingDirectory;
+        log.LogInformationMethod(new(), "wd3: {WorkingDirectory}", wd3);
+        Assert.Equal(wd1, wd3);
+    }
+
+    [SkippableFact]
+    public void GetAbsolutePath()
+    {
+        var dir = CreateTestWorkingDirectory();
+        var v = ConvertPath($"{dir.NameFull}/dir1/../dir1/./dir2/.");
+        var p = Client.GetAbsolutePath(v);
+        log.LogInformationMethod(new(), "{Original} --> {Absolute}", v, p);
+        Assert.Equal($"{dir.NameFull}/dir1/dir2", p);
+
+
     }
 
     [SkippableFact]
     public void Create_Directory_That_Already_Exists_Does_Not_Throw_Error()
     {
-        var dirName = HelperDeleteDirectoryIfExist();
+        var dirNameFull = GetTestDirectoryNameFull();
 
-        Client.CreateDirectory(dirName);
-        var dir = HelperGetDirectory(dirName);
+        if (Client.DirectoryExists(dirNameFull)) Client.DeleteDirectory(dirNameFull);
+        Client.CreateDirectory(dirNameFull);
+        var dir = GetDirectory(dirNameFull);
         Assert.NotNull(dir);
 
         // actual test
-        Client.CreateDirectory(dirName);
-        dir = HelperGetDirectory(dirName);
+        Client.CreateDirectory(dirNameFull);
+        dir = GetDirectory(dirNameFull);
         Assert.NotNull(dir);
     }
 
     [SkippableFact]
     public void PutFile_Auto_Create_SubDir()
     {
-        var dirName = CallerInfo.GetName()!;
-        var dir = HelperGetDirectory(dirName);
-        if (dir != null) Client.DeleteDirectory(dirName);
-        dir = HelperGetDirectory(dirName);
-        Assert.Null(dir);
+        var _ = CreateTestWorkingDirectory();
+        var dirSampleName = "sample";
 
-        Client.PutFile($"./{dirName}/myfile1.txt", ToBytes("My File 1"));
-        Client.PutFile($"{dirName}/myfile2.txt", ToBytes("My File 2"));
+        Client.PutFile(ConvertPath($"./{dirSampleName}/myfile1.txt"), ToBytes("My File 1"));
+        Client.PutFile(ConvertPath($"{dirSampleName}/myfile2.txt"), ToBytes("My File 2"));
 
-        var files = HelperGetFiles(dirName);
+        var files = GetFiles(dirSampleName);
 
         Assert.Equal(2, files.Length);
         Assert.Contains(files, o => o.Name == "myfile1.txt");
@@ -146,20 +185,19 @@ public abstract class FtpClientTests<T> : TestBase, IDisposable where T : FtpCli
     [SkippableFact]
     public void Directory_Operations()
     {
-        var dir = HelperGetDirectory("mydir");
-        if (dir != null) Client.DeleteDirectory("mydir");
+        var dir = CreateTestWorkingDirectory();
+        Assert.Equal(dir.NameFull, Client.WorkingDirectory);
 
-        Client.CreateDirectory("mydir");
-        dir = HelperGetDirectory("mydir");
-        Assert.NotNull(dir);
+        Client.WorkingDirectory = "..";
+        Assert.DoesNotContain(Client.DirectorySeparator + dir.Name, Client.WorkingDirectory);
 
-        Client.PutFile("./mydir/myfile1.txt", Constant.Encoding_UTF8.GetBytes("My File 1"));
-        Client.PutFile("mydir/myfile2.txt", Constant.Encoding_UTF8.GetBytes("My File 2"));
-        Client.WorkingDirectory = "./mydir";
+        Client.PutFile(ConvertPath($"./{dir.Name}/myfile1.txt"), Constant.Encoding_UTF8.GetBytes("My File 1"));
+        Client.PutFile(ConvertPath($"{dir.Name}/myfile2.txt"), Constant.Encoding_UTF8.GetBytes("My File 2"));
+        Client.WorkingDirectory = ConvertPath($"./{dir.Name}");
         Client.PutFile("myfile3.txt", Constant.Encoding_UTF8.GetBytes("My File 3"));
         Client.WorkingDirectory = "..";
 
-        var files = Client.ListObjects("mydir")
+        var files = Client.ListObjects(dir.Name)
             .Where(o => o.Type == FtpClientRemoteFileSystemObjectType.File)
             .Where(o => o.Name.In("myfile1.txt", "myfile2.txt", "myfile3.txt"))
             .ToList();
@@ -167,11 +205,18 @@ public abstract class FtpClientTests<T> : TestBase, IDisposable where T : FtpCli
         Assert.Equal(3, files.Count);
     }
 
+    #region Helpers
+
+    protected string ConvertPath(string unixPath) => unixPath.Replace("/", Client.DirectorySeparator);
+
+    protected static byte[] ToBytes(string str) => Constant.Encoding_UTF8.GetBytes(str);
+    protected static string ToString(byte[] bytes) => Constant.Encoding_UTF8.GetString(bytes);
+
     private void HelperDeleteFile(string filename)
     {
         try
         {
-            Client.DeleteFile(filename);
+            if (GetFile(filename) != null) Client.DeleteFile(filename);
         }
         catch (Exception)
         {
@@ -180,37 +225,50 @@ public abstract class FtpClientTests<T> : TestBase, IDisposable where T : FtpCli
         }
     }
 
-    private FtpClientRemoteFileSystemObject? HelperGetFile(string filename) =>
-        Client.ListObjects(null).FirstOrDefault(o => o.Type == FtpClientRemoteFileSystemObjectType.File && o.Name == filename);
+    private FtpClientRemoteFileSystemObject? GetFile(string fileName)
+    {
+        var o = Client.GetObject(fileName);
+        return o?.Type == FtpClientRemoteFileSystemObjectType.File ? o : null;
+    }
 
-    private FtpClientRemoteFileSystemObject[] HelperGetFiles(string? directoryName = null) =>
-        Client.ListObjects(directoryName).Where(o => o.Type == FtpClientRemoteFileSystemObjectType.File).ToArray();
+    private FtpClientRemoteFileSystemObject? GetDirectory(string directoryName)
+    {
+        var o = Client.GetObject(directoryName);
+        return o?.Type == FtpClientRemoteFileSystemObjectType.Directory ? o : null;
+    }
 
-    private FtpClientRemoteFileSystemObject[] HelperGetFilesRecursive(string? directoryName = null) =>
-        Client.ListObjectsRecursive(directoryName).Where(o => o.Type == FtpClientRemoteFileSystemObjectType.File).ToArray();
+    private FtpClientRemoteFileSystemObject[] GetFiles(string? directoryName = null) =>
+        Client.ListObjects(directoryName ?? Client.WorkingDirectory).Where(o => o.Type == FtpClientRemoteFileSystemObjectType.File).ToArray();
 
-    private FtpClientRemoteFileSystemObject? HelperGetDirectory(string directoryName) =>
-        Client.ListObjects(null).FirstOrDefault(o => o.Type == FtpClientRemoteFileSystemObjectType.Directory && o.Name == directoryName);
+    private FtpClientRemoteFileSystemObject[] GetFilesRecursive(string? directoryName = null) =>
+        Client.ListObjects(directoryName ?? Client.WorkingDirectory, true).Where(o => o.Type == FtpClientRemoteFileSystemObjectType.File).ToArray();
 
-    private string HelperDeleteDirectoryIfExist([CallerMemberName] string? directoryName = null)
+
+
+    //private string GetTestDirectoryName([CallerMemberName] string? directoryName = null) => directoryName.CheckNotNull();
+    private string GetTestDirectoryNameFull([CallerMemberName] string? directoryName = null) => Client.GetAbsolutePath(directoryName.CheckNotNull());
+    private FtpClientRemoteFileSystemObject CreateTestWorkingDirectory([CallerMemberName] string? directoryName = null)
     {
         Assert.NotNull(directoryName);
         directoryName = directoryName.CheckNotNull();
 
-        LogDisable();
-        var dir = HelperGetDirectory(directoryName);
-        if (dir != null) Client.DeleteDirectory(directoryName);
-        dir = HelperGetDirectory(directoryName);
-        LogEnable();
+        using (LogDisable())
+        {
+            directoryName = Client.GetAbsolutePath(directoryName);
+            if (Client.DirectoryExists(directoryName)) Client.DeleteDirectory(directoryName);
+            Client.CreateDirectory(directoryName);
+            var obj = Client.GetObject(directoryName);
+            Assert.NotNull(obj);
 
-        Assert.Null(dir);
-        return directoryName;
+            Client.WorkingDirectory = directoryName;
+            Assert.Empty(Client.ListObjects(directoryName, true));
+            return obj;
+        }
     }
-    public void Dispose()
-    {
-        var c = Client;
-        Client = null!;
-        // ReSharper disable once ConditionalAccessQualifierIsNonNullableAccordingToAPIContract
-        c?.Dispose();
-    }
+
+
+
+    #endregion Helpers
+
+
 }
